requirements.txt
fastapi==0.115.4
uvicorn==0.19.0
motor>=3.6
pydantic==1.10.7
passlib[bcrypt]==1.7.4
PyJWT==2.6.0
python-dotenv==0.21.1
python-json-logger==2.0.7
pymongo>=4.9
pydantic[email]
pytz
